requirements.txt
customtkinter==5.2.2
pyglet==2.1.11
pillow
